Employee project

+++ Employee/src/main/java/com/assessment/Employee/EmployeeContollers/EmployeeDetailsController.java
@@ -17,30 +17,41 @@ public class EmployeeDetailsController {
     @Autowired
     EmployeeDetailsService employeeDetailsService;
 
-    @PostMapping("/saveEmp")
+    @RequestMapping(value = "/saveEmp", method = RequestMethod.POST)
     public ResponseEntity<EmployeeDetailsEntity> createOrUpdateEmp(@RequestBody EmployeeDetailsEntity employeeDetailsEntity) {
-        EmployeeDetailsEntity emp = employeeDetailsService.createOrUpdateEmployee(employeeDetailsEntity);
+        EmployeeDetailsEntity emp = employeeDetailsService.createOrUpdateEmployee(employeeDetailsEntity, null);
         return new ResponseEntity<EmployeeDetailsEntity>(emp, new HttpHeaders(), HttpStatus.OK);
     }
 
-    @GetMapping("/getAllEmp")
+    @RequestMapping(value = "/updateEmp/{empId}", method = RequestMethod.PUT)
+    public ResponseEntity<EmployeeDetailsEntity> updateEmployee(@RequestBody EmployeeDetailsEntity employeeDetailsEntity, @PathVariable Long empId) {
+        EmployeeDetailsEntity emp = employeeDetailsService.createOrUpdateEmployee(employeeDetailsEntity, empId);
+        return new ResponseEntity<EmployeeDetailsEntity>(emp, new HttpHeaders(), HttpStatus.OK);
+    }
+
+    @RequestMapping(value = "/getEmp/{empId}", method = RequestMethod.GET)
+    public EmployeeDetailsEntity getEmployeeById(@PathVariable Long empId)
+            throws Exception {
+        EmployeeDetailsEntity employeeDetailsEntity = employeeDetailsService.getEmployeeDetailsById(empId);
+        if (null == employeeDetailsEntity)
+            throw new Exception("Could not find employee with Id " + employeeDetailsEntity.getEmpId());
+        return employeeDetailsEntity;
+    }
+
+    @RequestMapping(value = "/deleteEmp/{empId}", method = RequestMethod.DELETE)
+    public String deleteEmployee(@PathVariable Long empId) throws Exception {
+        EmployeeDetailsEntity employeeDetailsEntity = employeeDetailsService.getEmployeeDetailsById(empId);
+        if (null == employeeDetailsEntity)
+            throw new Exception("Could not find employee with Id " + employeeDetailsEntity.getEmpId());
+
+        employeeDetailsService.deleteEmployeebyId(empId);
+        return "deleted Successfull";
+    }
+
+    @RequestMapping(value = "/getAllEmp", method = RequestMethod.GET)
     public ResponseEntity<List<EmployeeDetailsEntity>> getAllEmployees() {
         List<EmployeeDetailsEntity> empList = employeeDetailsService.getAllEmployess();
         return new ResponseEntity<List<EmployeeDetailsEntity>>(empList, new HttpHeaders(), HttpStatus.OK);
-    }
-
-    @GetMapping("/{empId}")
-    public ResponseEntity<EmployeeDetailsEntity> getEmployeeById(@PathVariable("empId") Long empId)
-            throws Exception {
-        EmployeeDetailsEntity entity = employeeDetailsService.getEmployeeDetailsById(empId);
-
-        return new ResponseEntity<EmployeeDetailsEntity>(entity, new HttpHeaders(), HttpStatus.OK);
-    }
-
-    @DeleteMapping("/empId")
-    public HttpStatus deleteEmployee(@PathVariable("empId") Long empId) throws Exception {
-        boolean result = employeeDetailsService.deleteEmployeebyId(empId);
-        return HttpStatus.FORBIDDEN;
     }
 
 }
